[FEATURE] Set active TYPO3 extensions to env
To make it easier for composer managed installations,
it is now possible (again) to set active TYPO3 extensions
in the extra section of root composer.json

```
    "extra": {
        "helhum/typo3-console": {
            "active-typo3-extensions": [
				"belog",
				"beuser",
				"viewpage"
			]
        }
```

These will then again be exposed as environment variable
TYPO3_ACTIVE_FRAMEWORK_EXTENSIONS so that no change is needed in
TYPO3 console itself.

Also if the environment variable is already set, then this takes precedence
over the value set via this plugin.